+++ core/src/Transport.cc
@@ -139,7 +139,7 @@ void Surge::Transport::ArbitraryDataTransaction(const char *data, const size_t l
 
 void Surge::Transport::SafeRunLibuvCommand(std::function<void()> commandsToRun) {
     executingLibuvCommand = false;
-    m_loop->stop();
+    m_libuvCloser->send();
     commandsToRun();
     executingLibuvCommand = true;
 }
@@ -199,7 +199,7 @@ void Surge::Transport::AttachCallbacksToLibuv() {
     });
 
     m_libuvCloser->on<uvw::AsyncEvent>([this](const uvw::AsyncEvent &asyncEvent, uvw::AsyncHandle &asyncHandle) {
-        DEBUG("Transport thread: triggering stop request.");
+        DEBUG("Stopping transport thread.");
         m_loop->stop();
     });
 }